### 2018-06-11 #14 @gboeing
Paper is located at https://github.com/gboeing/pynamical/tree/master/paper
