### 2018-06-11 #14 @labarba
👋 @RobertTalbert — This one's for you, with all the beautiful math!

I'm not sure why our bot can't find the paper. Can you help us with that, @arfon?
