### 2018-06-12 #14 @labarba
@arfon This PDF build keeps failing. The error message says something about affiliation missing. I checked the paper metadata and the author affiliation is there. What could be going wrong?
