### 2018-06-13 #14 @labarba
Great, it compiled. Can you double check the bibliography? For some reason, the References list appears empty.
