### 2018-06-13 #14 @labarba
@gboeing Your final passage in the paper addresses the need: please add a heading to mark this "Statement of need." We'd like all papers to include this heading.
