### 2018-06-13 #14 @labarba
A few short items to add in the paper:
* has it been used in a classroom setting?
* how has it been used? how would someone else use it?
* what's the story behind it? how did the author come to write this?

Thanks!
