+++ paper/paper.bib
@@ -165,3 +165,15 @@
 	year = {1985},
 	pages = {364--395}
 }
+
+
+@inproceedings{tomida_matlab_2008,
+	address = {Perugia, Italy},
+	title = {Matlab {Toolbox} and {GUI} for {Analyzing} {One}-{Dimensional} {Chaotic} {Maps}},
+	isbn = {978-0-7695-3243-1},
+	doi = {10.1109/ICCSA.2008.7},
+	publisher = {IEEE},
+	author = {Tomida, Akemi Galvez},
+	year = {2008},
+	pages = {321--330}
+}
