### 2018-06-13 #15 @whedon
[REVIEW]: Pynamical: Model and Visualize Discrete Nonlinear Dynamical Systems, Chaos, and Fractals
**Submitting author:** @gboeing (<a href="http://orcid.org/0000-0003-1851-6411">Geoff Boeing</a>)
**Repository:** <a href="https://github.com/gboeing/pynamical" target ="_blank">https://github.com/gboeing/pynamical</a>
**Version:** 0.1.2

@drvinceknight, @sdross0, please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
