### 2018-06-13 #15 @drvinceknight
Proof looks good, I will complete the review before the weekend.
