### 2018-06-15 #15 @drvinceknight
I'm (really) impressed with the software and the paper is well written.

A couple of minor things that I have not checked off on my list and would like to run past @gboeing:

> Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

The documentation on github is great: I ran everything and it worked exactly as expected.

I noted that the documentation rendered by sphinx isn't complete. The autodoc of the api doesn't seem to have picked up the various methods/modules: https://pynamical.readthedocs.io/en/latest/pynamical.html#submodules

>  Tests: Are there automated tests or manual steps described so that the function of the software can be verified?

The test suite is currently just a test of run to completion of the various functions and not of expected output. I don't think this is something that should hinder publication at all but perhaps @gboeing you would want to think about adding some tests of expected output:

For example:

```
def test_phase_diagram_3d():
    pops = simulate(model=cubic_map, num_gens=200, rate_min=3.5, num_rates=30, num_discard=100)
    phase_diagram_3d(pops, xmin=-1, xmax=1, ymin=-1, ymax=1, zmin=-1, zmax=1, alpha=0.2, color='viridis', azim=330,
                     legend=True, save=False, folder=_img_folder, filename='')
    phase_diagram_3d(pops, xmin=-1, xmax=1, ymin=-1, ymax=1, zmin=-1, zmax=1, alpha=0.2, color='inferno', azim=330,
                     legend=True, remove_ticks=False, save=False, folder=_img_folder, filename='')

```

we could at least assert the type of `pops` (that it returns a dataframe) and possibly also the shape of the dataframe? Similarly for `phase_diagram_3d` (we can assert that we get a tuple, and the type of each element).

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

I didn't see these in the repo but perhaps I missed them?
