### 2018-06-15 4239bad
add community guidelines
