### 2018-06-15 689cea0
add community guidelines
