### 2018-06-15 #15 @gboeing
@drvinceknight thank you for these thoughtful recommendations. I have made each of these changes you suggested:

1. sphinx autodoc now renders all the docstrings properly: https://pynamical.readthedocs.io/en/latest/pynamical.html#submodules
1. new tests accordingly added in https://github.com/gboeing/pynamical/commit/861f9cec2608c0ffe7efdd588ae2f55878c3ca18
1. community contribution, issues, and support guidelines added in https://github.com/gboeing/pynamical/commit/689cea0ef4b30364091ffc8b2399e7e521fec45c
