### 2018-06-19 #15 @sdross0
Pros:
The software installs as advertised through both methods, and is able to successfully run all examples presented. Quick set-up and first examples worked out of the box. It runs very fast, which impresses me a lot. Very nice.

Cons:
From the docs, it's not clearly explained what class of problems this is designed to work for. Nonlinear dynamical systems could be any number of things (n-dimensional mappings, systems of ordinary differential equations, time-delay equations, etc.)

It seems to be designed primarily for 1D maps, that is, 1-dimensional discrete dynamical systems of the form

x_k+1 = f(x_k)

which isn't explicitly described in the documentation. I think a description of the kinds of models that the software has in mind will make it more successful. It would also help to know how to input your own models into the system.

Related to that, I worry the software might not generalize well to the other types of dynamical systems that would appear in an introductory course on dynamical systems (say, following Strogatz book), mostly because of the restriction to 1D maps.  Most such courses do include 1D maps at the beginning of the course, and perhaps that is where Pynamical could be most useful.
