### 2018-06-20 2a1cad1
clarify focus on 1D maps and beginning of intro course
