### 2018-06-20 be70559
clarify focus on 1D maps and link example of model defining
