+++ paper/paper.md
@@ -17,14 +17,14 @@ authors:
 affiliations:
   - name: University of California, Berkeley
     index: 1
-date: 15 May 2018
+date: 20 June 2018
 bibliography: paper.bib
 repository: https://github.com/gboeing/pynamical
 ---
 
 # Summary
 
-Pynamical is an educational Python package for introducing the modeling, simulation, and visualization of discrete nonlinear dynamical systems and chaos. Pynamical facilitates defining discrete one-dimensional nonlinear models as Python functions with just-in-time compilation for fast simulation. It comes packaged with the logistic map, the Singer map, and the cubic map predefined. The models may be run with a range of parameter values over a set of time steps, and the resulting numerical output is returned as a pandas DataFrame. Pynamical can then visualize this output in various ways, including with bifurcation diagrams [@may_simple_1976], two-dimensional phase diagrams [@packard_geometry_1980], three-dimensional phase diagrams, and cobweb plots [@hofstadter_mathematical_1985]. These visualizations enable simple qualitative assessments of system behavior including phase transitions [@feigenbaum_universal_1983], bifurcation points [@sander_many_2015], attractors and limit cycles [@grebogi_chaos_1987], basins of attraction [@sprott_classifying_2015], and fractals [@mandelbrot_how_1967; @mandelbrot_fractal_1983].
+Pynamical is an educational Python package for introducing the modeling, simulation, and visualization of discrete nonlinear dynamical systems and chaos, focusing on one-dimensional maps (such as the logistic map and the cubic map). Pynamical facilitates defining discrete one-dimensional nonlinear models as Python functions with just-in-time compilation for fast simulation. It comes packaged with the logistic map, the Singer map, and the cubic map predefined. The models may be run with a range of parameter values over a set of time steps, and the resulting numerical output is returned as a pandas DataFrame. Pynamical can then visualize this output in various ways, including with bifurcation diagrams [@may_simple_1976], two-dimensional phase diagrams [@packard_geometry_1980], three-dimensional phase diagrams, and cobweb plots [@hofstadter_mathematical_1985]. These visualizations enable simple qualitative assessments of system behavior including phase transitions [@feigenbaum_universal_1983], bifurcation points [@sander_many_2015], attractors and limit cycles [@grebogi_chaos_1987], basins of attraction [@sprott_classifying_2015], and fractals [@mandelbrot_how_1967; @mandelbrot_fractal_1983].
 
 Although most real-world systems are nonlinear dynamical systems, their mathematical analysis is notoriously difficult because they cannot be simply broken down into individual parts then recombined linearly [@strogatz_nonlinear_2014]. Instead, researchers have long relied on visualization techniques to make system behavior comprehensible [@alpigini_dynamical_2004; @layek_introduction_2015]. Such visualization is useful for exploring nonlinear time series data [@bradley_nonlinear_2015; @boeing_visual_2016]. Pynamical facilitates the modeling, visualization, and exploration of this rich nonlinear behavior, as demonstrated in Figures 1 and 2. Accordingly this package has various applications in research, engineering, and particularly pedagogy. For instance, it allows students to explore and model any well-defined discrete nonlinear dynamical system (with just one line of code) and then simulate its behavior over *n* generations in time (with just one more line of code). This facilitates easy experimentation in parameter space and sensitivity analysis. Moreover, Pynamical easily produces publication-quality visualizations to illustrate findings without having to reinvent the wheel with ad hoc algorithms to produce, reshape, and plot simulation data. Finally, it enables the animation of such systems, to explore and present their qualitative behavior and evolution.
 
@@ -32,7 +32,7 @@ The latest release of the software can be installed via `conda` or `pip` and ful
 
 # Statement of Need
 
[…]
