### 2018-06-20 471f11e
clarify focus on 1D maps and link example of model defining
