### 2018-06-20 22a3e7f
clarify focus on 1-D maps
