### 2018-06-20 #15 @gboeing
@sdross0 thank you for these thoughtful recommendations. I have made each of these changes you suggested:

  1. clarified the software's focus on 1-D maps in the documentation in https://github.com/gboeing/pynamical/commit/22a3e7fb70456891865837f3b86db8de538bbb09 and readme in https://github.com/gboeing/pynamical/commit/2a1cad13d71350401f6c91c20580e019c54eb9cf
  1. identified where this software could be most useful, accordingly, in the beginning of an intro course or demonstrating patterns to students in less technical fields, in https://github.com/gboeing/pynamical/commit/471f11ef2ffa0914901d0ffb21eb4e05e301cddb
  1. highlighted example of how to define/input your own model in https://github.com/gboeing/pynamical/commit/2a1cad13d71350401f6c91c20580e019c54eb9cf
