### 2018-06-20 #15 @labarba
@drvinceknight, @sdross0 — Thank you both for reviewing this submission to JOSE. Since you both have  ticked all items in your checklist, and made constructive comments that were addressed, all I need now is a confirmation on a comment below that you are ready to recommend acceptance.

This will be the first paper in JOSE and we're excited and grateful to you for being a part of this plunge into new a new model of scholarly publication!
