### 2018-06-20 #15 @drvinceknight
> now is a confirmation on a comment below that you are ready to recommend acceptance.

I confirm :+1:

>
> This will be the first paper in JOSE and we're excited and grateful to you for being a part of this plunge into a new model of scholarly publication!

Awesome: :+1:
