### 2018-06-20 #15 @sdross0
I recommend acceptance.
