### 2018-06-21 #15 @labarba
@gboeing -- your paper is now published in JOSE ‼️  Congratulations 🎉
You can now add the green happy JOSE button on your repository.

@drvinceknight, @sdross0 — Thank you again for your review 🙏
